archive/legacy/requirements_real_agents.txt
# Core FastAPI and web framework dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
jinja2==3.1.4

# Database dependencies
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# LangChain and AI dependencies
langchain==0.3.19
langchain-openai==0.2.14
langchain-google-genai==2.1.8
langgraph==0.2.60
langsmith==0.3.17

# Google AI dependencies
google-ai-generativelanguage==0.6.18
google-api-core[grpc]==2.24.1
google-auth==2.38.0

# Pydantic for data validation
pydantic==2.10.6
pydantic-settings==2.7.0

# HTTP client
httpx==0.28.1
requests==2.32.3

# Email validation
email-validator==2.2.0

# Calendar functionality
icalendar==6.0.1

# Environment and configuration
python-dotenv==1.0.1

# Utility libraries
typing-extensions==4.12.2
tenacity==9.0.0
jsonpatch==1.33
PyYAML==6.0.2
packaging==24.2
